Dev (#17)
dev

+++ pkg/storage/local_storage.go
@@ -7,6 +7,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/tomp332/p2p-agent/pkg"
 	"github.com/tomp332/p2p-agent/pkg/utils/configs"
+	"github.com/tomp332/p2p-agent/pkg/utils/types"
 	"io"
 	"math"
 	"os"
@@ -15,7 +16,7 @@ import (
 
 type LocalStorage struct {
 	options          *configs.LocalStorageConfig
-	maxFileSizeBytes float64
+	maxFileSizeBytes int64
 }
 
 func NewLocalStorage(options *configs.LocalStorageConfig) *LocalStorage {
@@ -31,23 +32,23 @@ func NewLocalStorage(options *configs.LocalStorageConfig) *LocalStorage {
 }
 
 func (s *LocalStorage) Initialize() error {
-	newPath := filepath.Join(s.options.RootDirectory, pkg.LocalStorageDefaultDir)
-	if _, err := os.Stat(newPath); errors.Is(err, os.ErrNotExist) {
-		if err = os.Mkdir(newPath, os.ModePerm); err != nil {
+	mainPath := filepath.Join(pkg.LocalStorageDefaultDir, s.options.RootDirectory)
+	if _, err := os.Stat(mainPath); errors.Is(err, os.ErrNotExist) {
+		if err = os.Mkdir(mainPath, os.ModePerm); err != nil {
 			log.Error().Msgf("Failed to create default storage directory, %s", err.Error())
 			return err
 		}
 	}
-	s.options.RootDirectory = newPath
+	s.options.RootDirectory = mainPath
 	if s.options.MaxStorageSize == 0 {
 		s.options.MaxStorageSize = math.MaxInt64
 	}
 	if s.options.MaxFileSize == 0 {
 		s.maxFileSizeBytes = 4 * 1024 * 1024 * 1024
 	} else {
-		s.maxFileSizeBytes = s.options.MaxFileSize * 1024 * 1024 * 1024
+		s.maxFileSizeBytes = int64(s.options.MaxFileSize * 1024 * 1024 * 1024)
 	}
-	if err := os.MkdirAll(newPath, os.ModePerm); err != nil {
+	if err := os.MkdirAll(mainPath, os.ModePerm); err != nil {
 		log.Error().
 			Str("rootDirectory", s.options.RootDirectory).
 			Msgf("Error initializing local file system storage. Failed to create root directory.")
@@ -55,13 +56,13 @@ func (s *LocalStorage) Initialize() error {
 	}
 	log.Debug().
 		Str("rootDirectory", s.options.RootDirectory).
-		Float64("maxFileSizeBytes", s.maxFileSizeBytes).
+		Int64("maxFileSizeBytes", s.maxFileSizeBytes).
 		Float64("MaxStorageSize", s.options.MaxStorageSize).
 		Msg("Initialized local storage successfully.")
 	return nil
 }
 
-func (s *LocalStorage) Put(ctx context.Context, fileID string, dataChan <-chan []byte) (float64, error) {
+func (s *LocalStorage) Put(ctx context.Context, fileID string, dataChan <-chan types.TransferChunkData) (int64, error) {
 	filePath := filepath.Join(s.options.RootDirectory, fileID)
 	file, err := os.Create(filePath)
 	if err != nil {
@@ -69,16 +70,19 @@ func (s *LocalStorage) Put(ctx context.Context, fileID string, dataChan <-chan [
 	}
 	defer file.Close()
 
-	var totalSize float64
-	done := make(chan error)
+	var totalSize int64
+	done := make(chan error, 1) // Buffer of 1 to handle the single error case
+
 	go func() {
-		for chunk := range dataChan {
-			n, err := file.Write(chunk)
-			if err != nil {
-				done <- err
+		defer close(done)
+		for data := range dataChan {
+			n, writeErr := file.Write(data.ChunkData)
+			if writeErr != nil {
+				log.Error().Err(writeErr).Msg("Error writing chunk to file")
+				done <- writeErr
 				return
 			}
-			totalSize += float64(n)
+			totalSize += int64(n)
 			if totalSize > s.maxFileSizeBytes {
 				done <- fmt.Errorf("file size exceeds the maximum allowed size of %.2f GB", s.options.MaxFileSize)
 				return
@@ -89,28 +93,21 @@ func (s *LocalStorage) Put(ctx context.Context, fileID string, dataChan <-chan [
 
 	select {
 	case <-ctx.Done():
-		os.Remove(filePath)
-		return 0, ctx.Err()
-	case err := <-done:
-		if err != nil {
-			log.Error().Msgf("Error uploading file: %v", err)
-			os.Remove(filePath)
-			return 0, err
+		// Context was canceled, but do not delete the file
+		return totalSize, ctx.Err()
+	case uploadErr := <-done:
+		if uploadErr != nil {
+			log.Error().Err(uploadErr).Msg("Error putting file in local storage")
+			os.Remove(filePath) // Attempt to remove the partially written file
+			return 0, uploadErr
 		}
 	}
+
 	log.Info().Str("fileId", fileID).Msg("File uploaded successfully.")
 	return totalSize, nil
 }
 
-func (s *LocalStorage) Search(fileID string) bool {
-	filePath := filepath.Join(s.options.RootDirectory, fileID)
-	if _, err := os.Stat(filePath); errors.Is(err, os.ErrNotExist) {
-		return false
-	}
-	return true
-}
-
-func (s *LocalStorage) Get(ctx context.Context, fileID string) (<-chan []byte, error) {
+func (s *LocalStorage) Get(ctx context.Context, fileID string) (<-chan types.TransferChunkData, error) {
 	filePath := filepath.Join(s.options.RootDirectory, fileID)
 	file, err := os.Open(filePath)
 	if err != nil {
@@ -119,7 +116,7 @@ func (s *LocalStorage) Get(ctx context.Context, fileID string) (<-chan []byte, e
 		}
 		return nil, err
 	}
-	dataChan := make(chan []byte)
+	dataChan := make(chan types.TransferChunkData)
 	go func() {
 		defer file.Close()
 		defer close(dataChan)
@@ -131,15 +128,18 @@ func (s *LocalStorage) Get(ctx context.Context, fileID string) (<-chan []byte, e
 				log.Info().Msg("Read file from storage canceled.")
 				return
 			default:
-				numBytesRead, err := file.Read(readBytes)
-				if err != nil {
-					if err != io.EOF {
-						log.Error().Err(err).Msg("Failed to read data from file")
+				numBytesRead, readErr := file.Read(readBytes)
+				if readErr != nil {
+					if readErr != io.EOF {
+						log.Error().Err(readErr).Msg("Failed to read data from file")
 					}
 					return
 				}
 				if numBytesRead > 0 {
-					dataChan <- readBytes[:numBytesRead]
+					dataChan <- types.TransferChunkData{
+						ChunkData: readBytes[:numBytesRead],
+						ChunkSize: int64(numBytesRead),
+					}
 				}
 				if numBytesRead < len(readBytes) {
 					log.Debug().Msg("Read file from storage successfully.")
